Delete project via wally-client
Removed authentication for now. It seemed pretty crude appending a well known password as a request param. Does that offer much? Going with nothing right now until something better can be implemented.

+++ lib/wally/application.rb
@@ -134,10 +134,9 @@ get '/projects/:project/?' do
 end
 
 delete '/projects/:project' do
-  error 403 unless authenticated?
-  project = Wally::Project.find_by_name(params[:project])
-  project.destroy
-  halt 201
+  if (project = Wally::Project.find_by_name(params[:project]))
+    project.destroy
+  end
 end
 
 get '/projects/:project/features/:feature/?' do
